+++ apps/desktop/main.go
@@ -7,6 +7,7 @@ import (
 	"embed"
 	"log"
 
+	"github.com/mow/mow/sdk/version"
 	"github.com/wailsapp/wails/v2"
 	"github.com/wailsapp/wails/v2/pkg/options"
 	"github.com/wailsapp/wails/v2/pkg/options/assetserver"
@@ -22,7 +23,7 @@ func main() {
 	}
 
 	err = wails.Run(&options.App{
-		Title:  "MOW Desktop",
+		Title:  "MOW Desktop " + version.Version,
 		Width:  1280,
 		Height: 800,
 		AssetServer: &assetserver.Options{